+++ protocol-ops/src/upgrade_verification/versions/v31/elements/set_new_version_upgrade/zksync_os.rs
@@ -1,0 +1,395 @@
+//! ZKsync OS `forceDeployAndUpgradeUniversal` payload verification.
+//!
+//! Owns the expected `UniversalContractUpgradeInfo[]` list (17 fixed-address
+//! entries with proxy-upgrade or unsafe-force-deployment shapes), the
+//! deployed-bytecode-info decoder (96-byte triple or 320-byte impl/proxy
+//! pair), the keccak-derived L2V31Upgrade delegate-address check, the ZKsync
+//! OS factory-dep bytecode list, and the ZKsync OS orchestrator wired from
+//! `ProposedUpgrade::verify_l2_protocol_upgrade_tx`.
+
+use alloy::primitives::{keccak256, Address, FixedBytes};
+use std::collections::HashMap;
+
+use crate::upgrade_verification::{
+    constants::{
+        GW_ASSET_TRACKER_ADDR, L2_ASSET_ROUTER_ADDR, L2_ASSET_TRACKER_ADDR,
+        L2_BASE_TOKEN_HOLDER_ADDR, L2_BASE_TOKEN_SYSTEM_CONTRACT_ADDR, L2_BRIDGEHUB_ADDR,
+        L2_CHAIN_ASSET_HANDLER_ADDR, L2_INTEROP_CENTER_ADDR, L2_INTEROP_HANDLER_ADDR,
+        L2_INTEROP_ROOT_STORAGE_ADDR, L2_MESSAGE_ROOT_ADDR, L2_MESSAGE_VERIFICATION_ADDR,
+        L2_NATIVE_TOKEN_VAULT_ADDR, L2_SYSTEM_CONTEXT_SYSTEM_CONTRACT_ADDR,
+        L2_SYSTEM_CONTRACT_PROXY_ADMIN_ADDR, L2_TO_L1_MESSENGER_SYSTEM_CONTRACT_ADDR,
+        L2_V31_UPGRADE_CONTRACT, L2_WRAPPED_BASE_TOKEN_IMPL_ADDR,
+    },
+    verifiers::{VerificationResult, Verifiers},
+};
+
+use super::{verify_l2_v31_upgrade_inner_calldata, IComplexUpgrader};
+
+/// ZKsync OS upgrade type — mirrors IComplexUpgrader.ContractUpgradeType.
+#[derive(Clone, Copy, PartialEq, Eq, Debug)]
+enum ZksyncOSUpgradeType {
+    SystemProxyUpgrade,
+    UnsafeForceDeployment,
+}
+
+struct ZksyncOSExpectedFd {
+    address: Address,
+    file: &'static str,
+    upgrade_type: ZksyncOSUpgradeType,
+}
+
+/// Expected v31 ZKsyncOS `UniversalContractUpgradeInfo[]` passed to
+/// `ComplexUpgrader.forceDeployAndUpgradeUniversal` — excludes the L2V31Upgrade delegate-target
+/// entry, which is validated separately by `verify_zksync_os_l2_v31_deployment`.
+fn expected_v31_zksync_os_force_deployments() -> Vec<ZksyncOSExpectedFd> {
+    macro_rules! proxy {
+        ($file:expr, $addr:expr) => {
+            ZksyncOSExpectedFd {
+                address: $addr,
+                file: $file,
+                upgrade_type: ZksyncOSUpgradeType::SystemProxyUpgrade,
+            }
+        };
+    }
+    macro_rules! unsafe_fd {
+        ($file:expr, $addr:expr) => {
+            ZksyncOSExpectedFd {
+                address: $addr,
+                file: $file,
+                upgrade_type: ZksyncOSUpgradeType::UnsafeForceDeployment,
+            }
+        };
+    }
+    vec![
+        // ── Fixed-address core contracts (getFixedAddressCoreContracts, 13 entries) ──
+        proxy!("l1-contracts/L2Bridgehub", L2_BRIDGEHUB_ADDR),
+        proxy!("l1-contracts/L2AssetRouter", L2_ASSET_ROUTER_ADDR),
+        proxy!(
+            "l1-contracts/L2NativeTokenVaultZKOS",
+            L2_NATIVE_TOKEN_VAULT_ADDR
+        ),
+        proxy!("l1-contracts/L2MessageRoot", L2_MESSAGE_ROOT_ADDR),
+        // L2WrappedBaseToken sits directly as the implementation — not behind a proxy.
+        unsafe_fd!(
+            "l1-contracts/L2WrappedBaseToken",
+            L2_WRAPPED_BASE_TOKEN_IMPL_ADDR
+        ),
+        proxy!(
+            "l1-contracts/L2MessageVerification",
+            L2_MESSAGE_VERIFICATION_ADDR
+        ),
+        proxy!(
+            "l1-contracts/L2ChainAssetHandler",
+            L2_CHAIN_ASSET_HANDLER_ADDR
+        ),
+        proxy!(
+            "l1-contracts/L2InteropRootStorage",
+            L2_INTEROP_ROOT_STORAGE_ADDR
+        ),
+        proxy!("l1-contracts/BaseTokenHolder", L2_BASE_TOKEN_HOLDER_ADDR),
+        proxy!("l1-contracts/L2AssetTracker", L2_ASSET_TRACKER_ADDR),
+        proxy!("l1-contracts/InteropCenter", L2_INTEROP_CENTER_ADDR),
+        proxy!("l1-contracts/InteropHandler", L2_INTEROP_HANDLER_ADDR),
+        proxy!("l1-contracts/GWAssetTracker", GW_ASSET_TRACKER_ADDR),
+        // ── ZKsync-OS system contracts (getZKsyncOSExtraSystemContracts, 3 entries) ──
+        proxy!("l1-contracts/L2BaseTokenZKOS", L2_BASE_TOKEN_SYSTEM_CONTRACT_ADDR),
+        proxy!("l1-contracts/L1MessengerZKOS", L2_TO_L1_MESSENGER_SYSTEM_CONTRACT_ADDR),
+        proxy!("l1-contracts/SystemContext", L2_SYSTEM_CONTEXT_SYSTEM_CONTRACT_ADDR),
+        // ── ProxyAdmin (_buildZKsyncOSProxyAdminEntry) ──
+        unsafe_fd!(
+            "l1-contracts/SystemContractProxyAdmin",
+            L2_SYSTEM_CONTRACT_PROXY_ADMIN_ADDR
+        ),
+    ]
+}
+
+/// Validate all entries of `UniversalContractUpgradeInfo[]` except the L2V31Upgrade delegate-target
+/// (which is already validated by `verify_zksync_os_l2_v31_deployment`).
+fn verify_v31_zksync_os_force_deployments(
+    verifiers: &Verifiers,
+    result: &mut VerificationResult,
+    deployments: &[IComplexUpgrader::UniversalContractUpgradeInfo],
+    delegate_to: Address,
+) {
+    let expected = expected_v31_zksync_os_force_deployments();
+    let mut expected_map: HashMap<Address, ZksyncOSExpectedFd> =
+        expected.into_iter().map(|e| (e.address, e)).collect();
+
+    for deployment in deployments {
+        // Skip the L2V31Upgrade delegate-target; already validated elsewhere.
+        if deployment.newAddress == delegate_to {
+            continue;
+        }
+
+        let addr = deployment.newAddress;
+        match expected_map.remove(&addr) {
+            None => {
+                result.report_error(&format!("Unexpected ZKsyncOS force deployment at {}", addr));
+            }
+            Some(expected_entry) => {
+                // Verify upgradeType.
+                let actual_upgrade_type = if deployment.upgradeType
+                    == IComplexUpgrader::ContractUpgradeType::ZKsyncOSSystemProxyUpgrade
+                {
+                    ZksyncOSUpgradeType::SystemProxyUpgrade
+                } else if deployment.upgradeType
+                    == IComplexUpgrader::ContractUpgradeType::ZKsyncOSUnsafeForceDeployment
+                {
+                    ZksyncOSUpgradeType::UnsafeForceDeployment
+                } else {
+                    result.report_error(&format!(
+                        "ZKsyncOS force deployment at {} ({}): unexpected upgradeType {:?}",
+                        addr, expected_entry.file, deployment.upgradeType
+                    ));
+                    continue;
+                };
+                if actual_upgrade_type != expected_entry.upgrade_type {
+                    result.report_error(&format!(
+                        "ZKsyncOS force deployment at {} ({}): upgradeType expected {:?}, got {:?}",
+                        addr, expected_entry.file, expected_entry.upgrade_type, actual_upgrade_type
+                    ));
+                }
+
+                // Verify deployedBytecodeInfo -> file.
+                verify_zksync_os_deployed_bytecode_info(
+                    verifiers,
+                    result,
+                    &deployment.deployedBytecodeInfo,
+                    expected_entry.file,
+                    &format!("{addr}"),
+                    expected_entry.upgrade_type,
+                );
+            }
+        }
+    }
+
+    let mut missing: Vec<_> = expected_map
+        .values()
+        .map(|e| format!("{} at {}", e.file, e.address))
+        .collect();
+    missing.sort();
+    for m in &missing {
+        result.report_error(&format!("Missing ZKsyncOS force deployment: {}", m));
+    }
+
+    if missing.is_empty() {
+        result.report_ok(
+            "All ZKsyncOS force deployments match the expected v31 list (excluding L2V31Upgrade delegate target)",
+        );
+    }
+}
+
+/// Verify the `deployedBytecodeInfo` of a ZKsyncOS force deployment entry maps to the expected file.
+///
+/// - `ZKsyncOSUnsafeForceDeployment`: 96-byte triple (blake2s | padding | keccak256); observable at [64..96].
+/// - `ZKsyncOSSystemProxyUpgrade`: abi.encode(implInfo_96, proxyInfo_96) = 320 bytes;
+///   impl observable (keccak256 of deployed impl bytecode) at [160..192].
+fn verify_zksync_os_deployed_bytecode_info(
+    verifiers: &Verifiers,
+    result: &mut VerificationResult,
+    bytecode_info: &[u8],
+    expected_file: &str,
+    addr_label: &str,
+    upgrade_type: ZksyncOSUpgradeType,
+) {
+    let (expected_len, observable_range) = match upgrade_type {
+        ZksyncOSUpgradeType::UnsafeForceDeployment => (96usize, 64..96),
+        ZksyncOSUpgradeType::SystemProxyUpgrade => (320usize, 160..192),
+    };
+
+    if bytecode_info.len() != expected_len {
+        result.report_error(&format!(
+            "ZKsyncOS force deployment at {addr_label} ({expected_file}): \
+             deployedBytecodeInfo length {} expected {expected_len}",
+            bytecode_info.len()
+        ));
+        return;
+    }
+
+    let observable = FixedBytes::<32>::from_slice(&bytecode_info[observable_range]);
+    if evm_deployed_bytecode_hash_matches_file(verifiers, &observable, expected_file) {
+        // ok — no noise on success to keep output readable
+    } else {
+        let actual_file = verifiers
+            .bytecode_verifier
+            .evm_deployed_bytecode_hash_to_file(&observable)
+            .cloned()
+            .unwrap_or_else(|| format!("unknown hash {observable}"));
+        result.report_error(&format!(
+            "ZKsyncOS force deployment at {addr_label}: expected file {expected_file}, \
+             observable hash maps to {actual_file}"
+        ));
+    }
+}
+
+/// ZKsync OS L2 factory-dep bytecode set. Mirrors
+/// `CoreOnGatewayHelper.getFullListOfFactoryDependencies(true, [L2V31Upgrade])`.
+pub(super) const EXPECTED_V31_ZKSYNC_OS_BYTECODES: &[&str] = &[
+    "l1-contracts/SystemContractProxy",
+    "l1-contracts/SystemContractProxyAdmin",
+    "l1-contracts/L2Bridgehub",
+    "l1-contracts/L2AssetRouter",
+    "l1-contracts/L2NativeTokenVaultZKOS",
+    "l1-contracts/L2MessageRoot",
+    "l1-contracts/L2WrappedBaseToken",
+    "l1-contracts/L2MessageVerification",
+    "l1-contracts/L2ChainAssetHandler",
+    "l1-contracts/L2InteropRootStorage",
+    "l1-contracts/BaseTokenHolder",
+    "l1-contracts/L2AssetTracker",
+    "l1-contracts/InteropCenter",
+    "l1-contracts/InteropHandler",
+    "l1-contracts/GWAssetTracker",
+    "l1-contracts/UpgradeableBeaconDeployer",
+    "l1-contracts/L2V31Upgrade",
+    "l1-contracts/L2BaseTokenZKOS",
+    "l1-contracts/L1MessengerZKOS",
+    "l1-contracts/SystemContext",
+];
+
+/// ZKsync OS orchestrator: walks the `UniversalContractUpgradeInfo[]`, validates the
+/// L2V31Upgrade delegate-target entry (derived address + bytecode info), then decodes
+/// the inner `IL2V31Upgrade.upgrade` calldata.
+pub(super) async fn verify_zksync_os_force_deploy_and_upgrade(
+    verifiers: &Verifiers,
+    result: &mut VerificationResult,
+    decoded: &IComplexUpgrader::forceDeployAndUpgradeUniversalCall,
+    expected_fixed_force_deployments_data: &str,
+) -> anyhow::Result<()> {
+    // Validate all expected force deployments (17 fixed entries; L2V31Upgrade delegate validated below).
+    verify_v31_zksync_os_force_deployments(
+        verifiers,
+        result,
+        &decoded._forceDeployments,
+        decoded._delegateTo,
+    );
+
+    // Validate the L2V31Upgrade delegate-target entry (1 unsafe force deployment at a derived address).
+    let mut matching_deployments = decoded
+        ._forceDeployments
+        .iter()
+        .filter(|deployment| deployment.newAddress == decoded._delegateTo);
+    match (matching_deployments.next(), matching_deployments.next()) {
+        (Some(deployment), None) => {
+            verify_zksync_os_l2_v31_deployment(verifiers, result, decoded._delegateTo, deployment);
+        }
+        (None, _) => result.report_error(&format!(
+            "ZKsync OS forceDeployAndUpgradeUniversal does not deploy delegate target {}",
+            decoded._delegateTo
+        )),
+        (Some(_), Some(_)) => result.report_error(&format!(
+            "ZKsync OS forceDeployAndUpgradeUniversal contains multiple deployments for delegate target {}",
+            decoded._delegateTo
+        )),
+    }
+
+    verify_l2_v31_upgrade_inner_calldata(
+        verifiers,
+        result,
+        &decoded._calldata,
+        true,
+        expected_fixed_force_deployments_data,
+    )
+    .await
+}
+
+fn verify_zksync_os_l2_v31_deployment(
+    verifiers: &Verifiers,
+    result: &mut VerificationResult,
+    delegate_to: Address,
+    deployment: &IComplexUpgrader::UniversalContractUpgradeInfo,
+) {
+    if deployment.upgradeType
+        != IComplexUpgrader::ContractUpgradeType::ZKsyncOSUnsafeForceDeployment
+    {
+        result.report_error(&format!(
+            "ZKsync OS L2V31Upgrade deployment must use ZKsyncOSUnsafeForceDeployment, got {:?}",
+            deployment.upgradeType
+        ));
+    }
+
+    let expected_delegate_to = generate_zksync_os_random_address(&deployment.deployedBytecodeInfo);
+    if delegate_to != expected_delegate_to {
+        result.report_error(&format!(
+            "ZKsync OS delegate target mismatch: expected derived address {}, got {}",
+            expected_delegate_to, delegate_to
+        ));
+    }
+
+    match zksync_os_bytecode_info_hashes(&deployment.deployedBytecodeInfo) {
+        Some((first_hash, observable_hash)) => {
+            if evm_deployed_bytecode_hash_matches_file(
+                verifiers,
+                &observable_hash,
+                L2_V31_UPGRADE_CONTRACT,
+            ) {
+                result.report_ok("ZKsync OS delegate deployment uses L2V31Upgrade bytecode info");
+            } else {
+                result.report_error(&format!(
+                    "ZKsync OS delegate bytecode info does not map to {}: blake={}, observable={}",
+                    L2_V31_UPGRADE_CONTRACT, first_hash, observable_hash
+                ));
+            }
+        }
+        None => result.report_error(&format!(
+            "ZKsync OS L2V31Upgrade bytecode info must be 96 bytes, got {}",
+            deployment.deployedBytecodeInfo.len()
+        )),
+    }
+}
+
+fn generate_zksync_os_random_address(bytecode_info: &[u8]) -> Address {
+    let mut preimage = Vec::with_capacity(32 + bytecode_info.len());
+    preimage.extend_from_slice(&[0u8; 32]);
+    preimage.extend_from_slice(bytecode_info);
+    let hash = keccak256(preimage);
+    Address::from_slice(&hash[12..])
+}
+
+fn zksync_os_bytecode_info_hashes(
+    bytecode_info: &[u8],
+) -> Option<(FixedBytes<32>, FixedBytes<32>)> {
+    if bytecode_info.len() != 96 {
+        return None;
+    }
+    Some((
+        FixedBytes::<32>::from_slice(&bytecode_info[0..32]),
+        FixedBytes::<32>::from_slice(&bytecode_info[64..96]),
+    ))
+}
+
+fn evm_deployed_bytecode_hash_matches_file(
+    verifiers: &Verifiers,
+    bytecode_hash: &FixedBytes<32>,
+    expected_file: &str,
+) -> bool {
+    verifiers
+        .bytecode_verifier
+        .evm_deployed_bytecode_hash_to_file(bytecode_hash)
+        .is_some_and(|file| file == expected_file)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn zksync_os_random_address_matches_helper_preimage_shape() {
+        let expected: Address = "0xbcd8f33061f2577d6118395e7b44ea21c7ef62e0"
+            .parse()
+            .unwrap();
+        assert_eq!(generate_zksync_os_random_address(&[1u8]), expected);
+    }
+
+    #[test]
+    fn zksync_os_bytecode_info_hashes_requires_abi_tuple_size() {
+        let mut bytecode_info = [0u8; 96];
+        bytecode_info[31] = 1;
+        bytecode_info[95] = 2;
+
+        let (first_hash, observable_hash) = zksync_os_bytecode_info_hashes(&bytecode_info).unwrap();
+        assert_eq!(first_hash[31], 1);
+        assert_eq!(observable_hash[31], 2);
+        assert!(zksync_os_bytecode_info_hashes(&bytecode_info[..95]).is_none());
+    }
+}
